Fix loop sync audit findings
